my_trend_app/requirements.txt
ib_insync
numpy
pandas
PyYAML
scikit-learn
ta
loguru
